requirements.txt
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.1
astroid==2.6.2
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
binaryornot==0.4.4
black==21.6b0
celery==4.4.6
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2
django-allauth==0.44.0
django-appconf==1.0.4
django-celery-beat==2.2.0
django-ckeditor==6.1.0
django-compressor==2.4.1
django-cors-headers==3.7.0
django-coverage-plugin==2.0.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeltranslation==0.17.3
django-redis==5.0.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
docutils==0.17.1
factory-boy==3.2.0
Faker==8.10.1
filelock==3.0.12
flake8==3.9.2
flake8-isort==4.0.0
future==0.18.2
h11==0.12.0
hiredis==1.1.0
httptools==0.2.0
identify==2.2.11
idna==3.2
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
kombu==4.6.11
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy==0.902
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.0.2
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.13.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.9.3
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-sugar==0.9.4
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
redis==3.5.3
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.18.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
vine==1.3.0
virtualenv==20.6.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==1.0.1
wrapt==1.12.1
